pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pogg"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Elizabeth C. Conrad", email = "lizcconrad@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "great_tables",
  "latextable",
  "networkx>=3.5",
  "pandas",
  "prettytable",
  "pydot",
  "pydelphin>=1.9.1",
  "pyyaml>=6.0.2",
  "tabulate>=0.9.0",
  "texttable"
]

[project.optional-dependencies]
data_conversion = [
  "swiplserver"
]

#[tool.hatch.envs.codev]
#pre-install-commands = [
#  "pip install -e ../pydelphin"
#]
#[tool.hatch.metadata]
#allow-direct-references = true

[project.urls]
Documentation = "https://github.com/lizcconrad/POGG#readme"
Issues = "https://github.com/lizcconrad/POGG/issues"
Source = "https://github.com/lizcconrad/POGG"

[tool.hatch.version]
path = "src/pogg/__about__.py"


[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0"
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/pogg tests}"

# workaround for pycharm for now ...?
[tool.hatch.envs.default]
type = "virtual"
path = "/Users/lizcconrad/.venvs/pogg_default"
features = [
  "data_conversion"
]

# docs environment
# inherits from default by default
[tool.hatch.envs.docs]
type = "virtual"
path = "/Users/lizcconrad/.venvs/pogg_docs"
dependencies = [
  "sphinx",
  "sphinx-rtd-theme",
  "sphinx-book-theme",
  "sphinx-autobuild",
  "sphinx-design",
  "myst-nb",
  "sphinx-autodoc2",
  "mystmd",
  "jupyterlab",
  "jupyterlab_myst",
  "sphinx-togglebutton"
]
[tool.hatch.envs.docs.scripts]
build = "make html"
clean = "make clean"
cleanbuild = "make clean & make html"
autobuild = "sphinx-autobuild source build/html"

# test environment
# inherits from default by default
[tool.hatch.envs.tests]
type = "virtual"
path = "/Users/lizcconrad/.venvs/pogg_tests"
# default dependencies for hatch-test but pycharm doesn't get along with hatch so i'm making it myself
dependencies = [
  "coverage-enable-subprocess==1.0",
  "coverage[toml]~=7.4",
  "pytest~=8.1",
  "pytest-mock~=3.12",
  "pytest-randomly~=3.15",
  "pytest-rerunfailures~=14.0",
  "pytest-xdist[psutil]~=3.5",
  "pytest-cases"
]

features = [
  "data_conversion"
]

[tool.coverage.run]
source_pkgs = ["pogg", "tests"]
branch = true
parallel = true
omit = [
  "src/pogg/__about__.py",
]

[tool.coverage.paths]
pogg = ["src/pogg", "*/pogg/src/pogg"]
tests = ["tests", "*/pogg/tests"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
